Conflicts:
	pom.xml
	src/main/webapp/WEB-INF/views/common/header.jsp
	src/main/webapp/WEB-INF/views/member/asideMain.jsp

+++ src/main/java/com/syi/project/model/notice/NoticeFileVO.java
@@ -10,7 +10,7 @@ public class NoticeFileVO {
 	private String fileOriginalName;
 	private String fileSavedName;
 	private String fileType;
-	private int fileSize;
+	private long fileSize;
 	private String filePath;
 	private Date fileRegDate;
 	private int fileNoticeNo;
